post CRUD endpoints completed with validations and middleware

+++ controllers/postController.js
@@ -1,6 +1,7 @@
 const { Post, User } = require('../models')
+const helper = require('../helpers/index')
 
-const NOT_FOUND = { message: 'user not found' }
+const NOT_FOUND = { message: 'post not found' }
 const SUCCESS = { message: "success" }
 
 const createPost = async (req, res) => {
@@ -8,7 +9,8 @@ const createPost = async (req, res) => {
         const post = await Post.create(req.body)
         res.status(201).send(post)
     } catch (error) {
-        res.send(error)
+        let status = error.name === "SequelizeUniqueConstraintError" ? 409 : 500
+        res.status(status).send(error.errors)
     }
 }
 
@@ -21,13 +23,16 @@ const getAllPosts = async (req, res) => {
         })
         res.status(200).send(posts)
     } catch (error) {
-        res.send(error)
+        res.status(500).send(error)
     }
 }
 
 const getPost = async (req, res) => {
     const { id } = req.params
     try {
+        let exist = await helper.resourceExists(Post, id)
+        if (!exist) return res.status(404).send(NOT_FOUND)
+
         const post = await Post.findByPk(id, {
             include: [{
                 model: User
@@ -35,22 +40,36 @@ const getPost = async (req, res) => {
         })
         res.status(200).send(post)
     } catch (error) {
-        res.send(error)
+        res.status(500).send(error)
     }
 }
 
 const updatePost = async (req, res) => {
     const { id } = req.params
+    try {
+        let exist = await helper.resourceExists(Post, id)
+        if (!exist) return res.status(404).send(NOT_FOUND)
 
-    await Post.update(req.body, { where: { id: id } })
-    res.send(SUCCESS)
+        await Post.update(req.body, { where: { id: id } })
+        res.status(200).send(SUCCESS)
+    } catch (error) {
+        res.status(500).send(error)
+    }
 }
 
 const deletePost = async (req, res) => {
     const { id } = req.params
-    await Post.destroy({ where: { id: id } })
-    res.send(SUCCESS)
+    try {
+        let exist = await helper.resourceExists(Post, id)
+        if (!exist) return res.status(404).send(NOT_FOUND)
+
+        await Post.destroy({ where: { id: id } })
+        res.send(SUCCESS)
+    } catch (error) {
+        res.status(500).send(error)
+    }
 }
+
 module.exports = {
     createPost,
     getAllPosts,
